Fix compilation errors on Linux
git-svn-id: http://svn.cern.ch/guest/lhcb/Online/trunk@95977 4525493e-7705-40b1-a816-d608a930855b

Former-commit-id: 0227d78a0e03cf054d4148f0c303d5389afc27ad

+++ Online/Online/OnlineKernel/OnlineKernel/RTL/screen.h
@@ -13,8 +13,8 @@ static inline void refresh()        {  }
 static inline WINDOW* initscreen()  { clrscr(); return 0; }
 static inline void printxy(int x, int y, const char* buff) {  ::gotoxy(x,y);  ::printf(buff); }
 static inline void print_char(int x, int y, int c) {  ::gotoxy(x,y); ::putchar(c); }
-static inline int  term_height() {  return   24; }
-static inline int  term_width()  {   return 132; }
+static inline size_t term_height() {  return   24; }
+static inline size_t term_width()  {   return 132; }
 
 #define NORMAL          0
 #define BOLD            0
